replace isAlive() in bgui, add nixt script and acquisition script
These are some fixes that resulted from a trip to Desy, where they
have rh9. Toko Hirono left a custom acquisition script at Desy which
I have committed.

The function isAlive() has been removed from the threading library,
and is_alive() is an alternative which works in python2/3.
The nixt script does nothing, or is a debug tool which is useful so
I decided to add it to github.

Also fixed the return codes of the wiener on/off scripts.